Change double evaluate to long

+++ AST/Expr.java
@@ -3,6 +3,6 @@ package AST;
 import java.util.Map;
 
 public interface Expr extends Node{
-    double eval(Map<String,Double> binding) throws EvalError;
+    long eval(Map<String,Long> binding) throws EvalError;
 }
