+++ AST/nearbyAST.java
@@ -12,7 +12,7 @@ public class nearbyAST implements Expr{
     }
 
     @Override
-    public double eval(Map<String, Double> binding) throws EvalError {
+    public long eval(Map<String, Long> binding) throws EvalError {
         return Command.instance().opponent();
     }
